refactor: extracted read request functions to `read-request.h`
