+++ include/sockets-polling.h
@@ -36,17 +36,7 @@
 #define GREY "\033[90m"
 #define BLUE "\033[0;34m"
 
-#define HEADER_END "\r\n\r\n"
-#define CRLF "\r\n"
-#define CRLF_LENGTH 2
-#define DOUBLE_CRLF_LENGTH 4
 
-enum readReturnVal {
-    RECV_OK = 0,
-    RECV_AGAIN = -1,
-    RECV_CLOSED = -2,
-	REQUEST_ERR = -3,
-};
 
 // Setup Listening Socket
 void setupListeningSocket(std::vector<struct pollfd>& pfds, std::vector<int>& listeners, Server& server);
@@ -55,6 +45,7 @@ bool isListener(std::vector<int>& listeners, int fd);
 Client* findClientByFd(std::vector<Client>& clients, int fd);
 
 void acceptClient(std::vector<struct pollfd>& pfds, std::vector<Client>& clients, int listener);
+
 // handle poll events
 void handlePollIn(std::map< std::pair<std::string, std::string> , std::vector<Server> >& servers,
                     struct pollfd& pfd, Client& client);
@@ -63,16 +54,15 @@ void handlePollHup(Client& client);
 void handlePollErr(Client& client);
 
 
-// Read Request Utils
-int readRequestHeader(Client &client, std::string& headerStr, const size_t bufferSize);
-int readRequestBody(Client &client, std::string& bodyStr, const size_t bufferSize, const size_t maxSize);
-int receiveClientRequest(Client &client, std::map< std::pair<std::string, std::string> , std::vector<Server> >& servers);
-int readByChunkedEncoding(Client &client, std::string& bodyStr, const size_t bufferSize, const size_t maxSize);
+// // Read Request Utils
+// int readRequestHeader(Client &client, std::string& headerStr, const size_t bufferSize);
+// int readRequestBody(Client &client, std::string& bodyStr, const size_t bufferSize, const size_t maxSize);
+// int receiveClientRequest(Client &client, std::map< std::pair<std::string, std::string> , std::vector<Server> >& servers);
+// int readByChunkedEncoding(Client &client, std::string& bodyStr, const size_t bufferSize, const size_t maxSize);
 
 // Utils
 int  setNonBlocking(int fd);
 std::vector<Client>::iterator disconnectClient(std::vector<Client>& clients, std::vector<Client>::iterator &clientIt, std::vector<struct pollfd>& pfds);
-int readFromSocket(Client &client, int bufferSize);
 void clearDisconnectedClients(std::vector<Client>& clients, std::vector<struct pollfd>& pfds);
 
 // utils2
