+++ src/read-request/receive-client-request.cpp
@@ -1,5 +1,6 @@
-#include "../../include/sockets-polling.h"
-#include "../../include/http-request.h"
+#include "sockets-polling.h"
+#include "http-request.h"
+#include "read-request.h"
 
 int handleParsingError(const HttpException& e, HttpResponse& response, Client& client)
 {
